Foord-python/requirements.txt
awsiotsdk==1.11.5
python-dotenv==0.20.0
